Expose way to get Context
